Add initial audit trail implementation

+++ backend/src/access-control/services/accessControlService.ts
@@ -1,6 +1,8 @@
 import { Request, Response } from "express";
-import { createDepartment, deleteDepartment, getDepartments, createAdmin, getAdmins, deleteAdmin, createUser, deleteUser, getUsers, getUser, getDepartment } from '../services/dbService'
+import { createDepartment, deleteDepartment, getDepartments, createAdmin, getAdmins, deleteAdmin, createUser, deleteUser, getUsers, getUser, getDepartment, getDepartmentById, getAdmin } from '../services/dbService'
 import { getUserDataByEmail } from './ringcentralService'
+import { addAuditTrailItem } from "../../audit-trail/services/dbService";
+import { AuditTrailItem } from "../../audit-trail/interface/AuditTrailItem";
 
 // Departments -----------------------------------------------------------------
 
@@ -26,22 +28,56 @@ export const processCreateDepartmentRequest = async (req: Request, res: Response
         return
     }
 
+    const auditItem: AuditTrailItem = {
+        action: `Added ${name} to the department whitelist`,
+        initiator: addedByName,
+        tool: 'Management Console',
+        type: 'access'
+    }
+    addAuditTrailItem(auditItem)
+
     res.status(200).send()
 }
 
 export const processDeleteDepartmentRequest = async (req: Request, res: Response) => {
     const id = parseInt(req.params.id)
+    const addedByName = req.headers.addedByName
+    const addedByEmail = req.headers.addedByEmail
 
     if (!id || isNaN(id)) {
         res.status(400).send()
         return
     }
 
-    await deleteDepartment(id)
+    const department = await getDepartmentById(id)
+
+    if (!department) {
+        res.status(404).send()
+        return
+    }
+
+    const success = await deleteDepartment(id)
+
+    if (!success) {
+        res.status(500).send()
+        return
+    }
+
+    const auditItem: AuditTrailItem = {
+        action: `Removed ${department.name} from the department whitelist`,
+        initiator: `${addedByName}`,
+        tool: 'Management Console',
+        type: 'access'
+    }
+    addAuditTrailItem(auditItem)
+
     res.status(200).send()
 }
 
 export const isDepartmentWhiteListed = async (name: string) => {
+    if (!name || typeof name !== 'string') {
+        return false
+    }
     const department = await getDepartment(name)
     return department !== null
 }
@@ -73,18 +109,48 @@ export const processCreateAdminRequest = async (req: Request, res: Response) => 
         return
     }
 
+    const auditItem: AuditTrailItem = {
+        action: `Added ${userData.name} as an admin`,
+        initiator: addedByName,
+        tool: 'Management Console',
+        type: 'access'
+    }
+    addAuditTrailItem(auditItem)
+
     res.status(200).send()
 }
 
 export const processDeleteAdminRequest = async (req: Request, res: Response) => {
     const id = parseInt(req.params.id)
+    const addedByName = req.headers.addedByName
 
     if (!id || isNaN(id)) {
         res.status(400).send()
         return
     }
 
-    await deleteAdmin(id)
+    const admin = await getAdmin(id)
+
+    if (!admin) {
+        res.status(404).send()
+        return
+    }
+
+    const success = await deleteAdmin(id)
+
+    if (!success) {
+        res.status(500).send()
+        return
+    }
+
+    const auditItem: AuditTrailItem = {
+        action: `Removed ${admin.name} as an admin`,
+        initiator: `${addedByName}`,
+        tool: 'Management Console',
+        type: 'access'
+    }
+    addAuditTrailItem(auditItem)
+
     res.status(200).send()
 }
 
@@ -120,18 +186,48 @@ export const processCreateUserRequest = async (req: Request, res: Response) => {
         return
     }
 
+    const auditItem: AuditTrailItem = {
+        action: `Added ${userData.name} to the user whitelist`,
+        initiator: addedByName,
+        tool: 'Management Console',
+        type: 'access'
+    }
+    addAuditTrailItem(auditItem)
+
     res.status(200).send()
 }
 
 export const processDeleteUserRequest = async (req: Request, res: Response) => {
     const id = parseInt(req.params.id)
+    const addedByName = req.headers.addedByName
 
     if (!id || isNaN(id)) {
         res.status(400).send()
         return
     }
 
-    await deleteUser(id)
+    const user = await getUser(id)
+
+    if (!user) {
+        res.status(404).send()
+        return
+    }
+
+    const success = await deleteUser(id)
+
+    if (!success) {
+        res.status(500).send()
+        return
+    }
+
+    const auditItem: AuditTrailItem = {
+        action: `Removed ${user.name} from the user whitelist`,
+        initiator: `${addedByName}`,
+        tool: 'Management Console',
+        type: 'access'
+    }
+    addAuditTrailItem(auditItem)
+
     res.status(200).send()
 }
